backend/requirements.txt
flask
flask-cors
kafka-python
scikit-learn
pandas
numpy
sqlalchemy
faker
python-dotenv
joblib
requests
